# core/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CauseFlow API
  description: |
    CauseFlow — AI SRE platform. REST API for incident ingestion, triage,
    investigation, remediation, memory/knowledge, integrations, billing,
    tenant/user/team management, notifications, audit, widget/portal, and
    skills.

    Architecture: Modular monolith (Modlito) + Clean Architecture.
    Stack: TypeScript + Hono + ElectroDB (DynamoDB single-table) + Redis + SQS.

    ## Authentication
    Most dashboard endpoints use `bearerAuth` (Clerk JWT).
    Ingestion webhook endpoints accept either a shared webhook secret or a
    tenant API key (`X-API-Key`). Widget endpoints use `X-API-Key`.
    Stripe, Clerk and Composio webhooks verify provider signatures and are
    mounted as public.
  version: 0.1.0
  contact:
    name: CauseFlow
servers:
  - url: https://api.causeflow.io
    description: Production
  - url: https://staging-api.causeflow.io
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/incidents:
    get:
      tags: [Incident]
      summary: List incidents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - { in: query, name: severity, schema: { type: string, enum: [critical, high, medium, low, info] } }
        - { in: query, name: status, schema: { type: string, enum: [open, triaging, investigating, awaiting_approval, remediating, resolved, closed] } }
        - { in: query, name: sort, schema: { type: string, enum: [asc, desc] } }
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Incident' } }
                  cursor: { type: string, nullable: true }
  /v1/incidents/chat:
    post:
      tags: [Incident]
      summary: Create incident manually via chat (operator/admin/owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description]
              properties:
                title: { type: string, minLength: 5 }
                description: { type: string, minLength: 10 }
                severity: { type: string, enum: [critical, high, medium, low, info] }
                suggestedAgents: { type: array, items: { type: string } }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  incidentId: { type: string }
                  status: { type: string }
                  message: { type: string }
        '402': { description: Investigation quota exceeded }
  /v1/incidents/{id}:
    get:
      tags: [Incident]
      summary: Get an incident by id
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
    patch:
      tags: [Incident]
      summary: Update incident status (operator/admin/owner)
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, triaging, investigating, awaiting_approval, remediating, resolved, closed] }
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }

  # ─────────────────────────── Ingestion: Webhooks ───────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT. Tenant context derived from `o.id` claim.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant-scoped API key for programmatic and widget access.
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared webhook secret for ingestion endpoints.
    webhookSignature:
      type: apiKey
      in: header
      name: webhook-signature
      description: HMAC signature header used by Composio webhooks.
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: Stripe webhook signature header.
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix signature for Clerk webhooks.
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
...
  schemas:
...
    Incident:
      type: object
      properties:
        incidentId: { type: string }
        tenantId: { type: string }
        title: { type: string }
        description: { type: string }
        severity: { type: string, enum: [critical, high, medium, low, info] }
        status: { type: string, enum: [open, triaging, investigating, awaiting_approval, remediating, resolved, closed] }
        source: { type: string }
        externalId: { type: string }
        suggestedAgents: { type: array, items: { type: string } }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...